Move TOTP CRUD actions to TotpsController
Extracted from MultifactorAuthsController
