+++ app/controllers/multifactor_auths_controller.rb
@@ -4,45 +4,15 @@ class MultifactorAuthsController < ApplicationController
   include WebauthnVerifiable
 
   before_action :redirect_to_signin, unless: :signed_in?
-  before_action :require_totp_disabled, only: %i[new create]
   before_action :require_mfa_enabled, only: %i[update otp_update]
-  before_action :require_totp_enabled, only: :destroy
-  before_action :seed_and_expire, only: :create
   before_action :find_mfa_user, only: %i[update otp_update webauthn_update]
   before_action :require_mfa, only: %i[update]
   before_action :validate_otp, only: %i[otp_update]
   before_action :require_webauthn_enabled, only: %i[webauthn_update]
   before_action :validate_webauthn, only: %i[webauthn_update]
-  before_action :disable_cache, only: %i[new recovery]
+  before_action :disable_cache, only: %i[recovery]
   after_action :delete_mfa_session, only: %i[otp_update webauthn_update]
   helper_method :issuer
-
-  def new
-    @seed = ROTP::Base32.random_base32
-    session[:totp_seed] = @seed
-    session[:totp_seed_expire] = Gemcutter::MFA_KEY_EXPIRY.from_now.utc.to_i
-    text = ROTP::TOTP.new(@seed, issuer: issuer).provisioning_uri(current_user.email)
-    @qrcode_svg = RQRCode::QRCode.new(text, level: :l).as_svg(module_size: 6)
-  end
-
-  def create
-    current_user.verify_and_enable_totp!(@seed, :ui_and_api, otp_param, @expire)
-    if current_user.errors.any?
-      flash[:error] = current_user.errors[:base].join
-      redirect_to edit_settings_url
-    else
-      flash[:success] = t(".success")
-      @continue_path = session.fetch("mfa_redirect_uri", edit_settings_path)
-
-      if current_user.mfa_device_count_one?
-        session[:show_recovery_codes] = current_user.new_mfa_recovery_codes
-        redirect_to recovery_multifactor_auth_path
-      else
-        redirect_to @continue_path
-        session.delete("mfa_redirect_uri")
-      end
-    end
-  end
 
   # not possible to arrive here because of require_mfa_enabled + require_mfa, but it must stay for rails to be happy.
   def update
@@ -54,18 +24,6 @@ class MultifactorAuthsController < ApplicationController
 
   def webauthn_update
     update_level_and_redirect
-  end
-
-  def destroy
-    if current_user.ui_mfa_verified?(otp_param)
-      flash[:success] = t(".success")
-      current_user.disable_totp!
-      redirect_to session.fetch("mfa_redirect_uri", edit_settings_path)
-      session.delete("mfa_redirect_uri")
-    else
-      flash[:error] = t("multifactor_auths.incorrect_otp")
-      redirect_to edit_settings_path
-    end
   end
 
   def recovery
@@ -83,10 +41,6 @@ class MultifactorAuthsController < ApplicationController
 
   private
 
-  def otp_param
-    params.permit(:otp).fetch(:otp, "")
-  end
-
   def level_param
     params.permit(:level).require(:level)
   end
@@ -95,23 +49,9 @@ class MultifactorAuthsController < ApplicationController
     request.host || "rubygems.org"
   end
 
-  def require_totp_disabled
-    return if current_user.totp_disabled?
-    flash[:error] = t("multifactor_auths.require_totp_disabled", host: Gemcutter::HOST_DISPLAY)
-    redirect_to edit_settings_path
-  end
-
   def require_mfa_enabled
     return if current_user.mfa_enabled?
     flash[:error] = t("multifactor_auths.require_mfa_enabled")
-    redirect_to edit_settings_path
-  end
-
-  def require_totp_enabled
-    return if current_user.totp_enabled?
-
-    flash[:error] = t("multifactor_auths.require_totp_enabled")
-    delete_mfa_session
     redirect_to edit_settings_path
   end
 
@@ -121,11 +61,6 @@ class MultifactorAuthsController < ApplicationController
     flash[:error] = t("multifactor_auths.require_webauthn_enabled")
     delete_mfa_session
     redirect_to edit_settings_path
-  end
-
-  def seed_and_expire
-    @seed = session.delete(:totp_seed)
-    @expire = Time.at(session.delete(:totp_seed_expire) || 0).utc
   end
 
   def update_level_and_redirect
