+++ app/models/user/with_private_fields.rb
@@ -11,7 +11,7 @@ class User::WithPrivateFields < User
   def mfa_warning
     if mfa_recommended_not_yet_enabled?
       "[WARNING] For protection of your account and gems, we encourage you to set up multi-factor authentication " \
-        "at https://rubygems.org/multifactor_auth/new. Your account will be required to have MFA enabled in the future."
+        "at https://rubygems.org/totp/new. Your account will be required to have MFA enabled in the future."
     elsif mfa_recommended_weak_level_enabled?
       "[WARNING] For protection of your account and gems, we encourage you to change your multi-factor authentication " \
         "level to 'UI and gem signin' or 'UI and API' at https://rubygems.org/settings/edit. " \
