+++ app/views/totps/new.html.erb
@@ -14,7 +14,7 @@
   </div>
 </div>
 
-<%= form_tag multifactor_auth_path, method: :post do %>
+<%= form_tag totp_path, method: :post do %>
   <div class="text_field">
     <%= label_tag :otp, 'OTP code', class: 'form__label' %>
     <p class='form__field__instructions'><%= t '.otp_prompt' %></p>
