+++ config/locales/es.yml
@@ -474,14 +474,9 @@ es:
     update:
       failure:
   multifactor_auths:
-    incorrect_otp: Tu código OTP no es correcto.
     session_expired: Ha expirado tu sesión en la página de acceso.
-    require_totp_disabled: La autenticación de múltiples factores basada en OTP ya
-      está activa. Para reconfigurarla debes primero eliminarla.
     require_mfa_enabled: No se ha activado la autenticación de múltiples factores.
       Primero tienes que activarla.
-    require_totp_enabled: No tienes aplicación de autenticación activa. Debes activarla
-      primero.
     require_webauthn_enabled: No tienes ningún dispositivo de seguridad activado.
       Primero debes asociar un dispositivo a tu cuenta.
     setup_required_html: Por la seguridad de tu cuenta y de tus gemas se te requiere
@@ -501,21 +496,6 @@ es:
     setup_webauthn_html: "\U0001F389 ¡Ahora soportamos dispositivos de seguridad!
       Aumenta la seguridad de tu cuenta <a href=\"/settings/edit#security-device\">configurando</a>
       un nuevo dispositivo."
-    new:
-      title: Activando autenticación de múltiples factores
-      scan_prompt: Por favor escanea el código QR con tu aplicación de autenticación.
-        Si no puedes escanear el código, agrega manualmente la información siguiente
-        a tu aplicación.
-      otp_prompt: Escribe el código de la aplicación de autenticación para continuar.
-      confirm: He mantenido seguros mis códigos de recuperación.
-      enable: Activar
-      account: 'Cuenta: %{account}'
-      key: 'Clave: %{key}'
-      time_based: 'Basado en tiempo: Sí'
-    create:
-      qrcode_expired: El código QR y la clave han vencido. Por favor intenta otra
-        vez registrar un nuevo dispositivo.
-      success: Has activado con éxito la autenticación de múltiples factores.
     recovery:
       copied: "[ copiado ]"
       continue: Continuar
@@ -527,8 +507,6 @@ es:
         tu autenticación de múltiples factores si pierdes tu dispositivo. Cada código
         se puede usar una sola vez.
       already_generated: Ya deberías haber guardado tus códigos de recuperación.
-    destroy:
-      success: Has desactivado exitosamente la autenticación de múltiples factores.
     update:
       invalid_level: Nivel de AMF inválido.
       success: Has actualizado exitosamente la autenticación de múltiples factores.
@@ -544,6 +522,29 @@ es:
         a tu dispositivo de seguridad o de autenticación de múltiples factores.
       security_device: Dispositivo de seguridad
       verify_code: Verificar código
+  totps:
+    incorrect_otp: Tu código OTP no es correcto.
+    require_totp_disabled: La autenticación de múltiples factores basada en OTP ya
+      está activa. Para reconfigurarla debes primero eliminarla.
+    require_totp_enabled: No tienes aplicación de autenticación activa. Debes activarla
+      primero.
+    new:
+      title: Activando autenticación de múltiples factores
+      scan_prompt: Por favor escanea el código QR con tu aplicación de autenticación.
+        Si no puedes escanear el código, agrega manualmente la información siguiente
+        a tu aplicación.
+      otp_prompt: Escribe el código de la aplicación de autenticación para continuar.
+      confirm: He mantenido seguros mis códigos de recuperación.
+      enable: Activar
+      account: 'Cuenta: %{account}'
+      key: 'Clave: %{key}'
+      time_based: 'Basado en tiempo: Sí'
+    create:
+      qrcode_expired: El código QR y la clave han vencido. Por favor intenta otra
+        vez registrar un nuevo dispositivo.
+      success: Has activado con éxito la autenticación de múltiples factores.
+    destroy:
+      success: Has desactivado exitosamente la autenticación de múltiples factores.
   notifiers:
     update:
       success: Has actualizado exitosamente la configuración de tus notificaciones
