+++ config/locales/pt-BR.yml
@@ -435,29 +435,14 @@ pt-BR:
     update:
       failure:
   multifactor_auths:
-    incorrect_otp:
     session_expired:
-    require_totp_disabled:
     require_mfa_enabled:
-    require_totp_enabled:
     require_webauthn_enabled:
     setup_required_html:
     setup_recommended:
     strong_mfa_level_required_html:
     strong_mfa_level_recommended:
     setup_webauthn_html:
-    new:
-      title:
-      scan_prompt:
-      otp_prompt:
-      confirm:
-      enable:
-      account:
-      key:
-      time_based:
-    create:
-      qrcode_expired:
-      success:
     recovery:
       copied:
       continue:
@@ -466,8 +451,6 @@ pt-BR:
       saved:
       note_html:
       already_generated:
-    destroy:
-      success:
     update:
       invalid_level:
       success:
@@ -480,6 +463,24 @@ pt-BR:
       recovery_code_html:
       security_device:
       verify_code:
+  totps:
+    incorrect_otp:
+    require_totp_disabled:
+    require_totp_enabled:
+    new:
+      title:
+      scan_prompt:
+      otp_prompt:
+      confirm:
+      enable:
+      account:
+      key:
+      time_based:
+    create:
+      qrcode_expired:
+      success:
+    destroy:
+      success:
   notifiers:
     update:
       success:
