+++ test/functional/api/v1/api_keys_controller_test.rb
@@ -471,7 +471,7 @@ class Api::V1::ApiKeysControllerTest < ActionController::TestCase
           assert_response 403
           mfa_error = <<~ERROR.chomp
             [ERROR] For protection of your account and your gems, you are required to set up multi-factor authentication \
-            at https://rubygems.org/multifactor_auth/new.
+            at https://rubygems.org/totp/new.
 
             Please read our blog post for more details (https://blog.rubygems.org/2022/08/15/requiring-mfa-on-popular-gems.html).
           ERROR
