+++ test/functional/api/v1/rubygems_controller_test.rb
@@ -562,7 +562,7 @@ class Api::V1::RubygemsControllerTest < ActionController::TestCase
         should "show error message" do
           mfa_error = <<~ERROR.chomp
             [ERROR] For protection of your account and your gems, you are required to set up multi-factor authentication \
-            at https://rubygems.org/multifactor_auth/new.
+            at https://rubygems.org/totp/new.
 
             Please read our blog post for more details (https://blog.rubygems.org/2022/08/15/requiring-mfa-on-popular-gems.html).
           ERROR
@@ -634,7 +634,7 @@ class Api::V1::RubygemsControllerTest < ActionController::TestCase
 
 
             [WARNING] For protection of your account and gems, we encourage you to set up multi-factor authentication \
-            at https://rubygems.org/multifactor_auth/new. Your account will be required to have MFA enabled in the future.
+            at https://rubygems.org/totp/new. Your account will be required to have MFA enabled in the future.
           WARN
 
           assert_includes @response.body, mfa_warning
